requirements.txt
flask==2.3.3
Werkzeug==2.3.7
python-docx==0.8.11
PyPDF2==3.0.1
PyMuPDF==1.22.5
google-generativeai==0.3.1